ex1-7a edits

+++ ex1-7a-count-chars/main.c
@@ -2,9 +2,9 @@
 
 int main(void)
 {
-	long nc;
+	double nc;
 
 	for (nc = 0; getchar() != EOF; ++nc)
 		;
-	printf("%ld\n", nc);
+	printf("%.0f\n", nc);
 }
